+++ x/wasm/binding/cw_struct/cw_msg.go
@@ -21,6 +21,7 @@ type BindingMsg struct {
 	InsuranceFundWithdraw *InsuranceFundWithdraw `json:"insurance_fund_withdraw,omitempty"`
 	PegShift              *PegShift              `json:"peg_shift,omitempty"`
 	DepthShift            *DepthShift            `json:"depth_shift,omitempty"`
+	SetMarketEnabled      *SetMarketEnabled      `json:"set_market_enabled,omitempty"`
 
 	EditOracleParams *EditOracleParams `json:"edit_oracle_params,omitempty"`
 }
@@ -91,4 +92,9 @@ type EditOracleParams struct {
 type InsuranceFundWithdraw struct {
 	Amount sdk.Int `json:"amount"`
 	To     string  `json:"to"`
+}
+
+type SetMarketEnabled struct {
+	Pair    string `json:"pair"`
+	Enabled bool   `json:"enabled"`
 }